backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.31.0
pydantic==2.9.0
roboflow==1.1.32
ultralytics==8.3.25
torch
torchvision
matplotlib
numpy
scikit-learn
reportlab
pillow
